# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ApplyFriendRequest:
    properties:
      avatar:
        description: 网站头像/logo
        maxLength: 255
        type: string
      description:
        description: 网站描述
        maxLength: 500
        type: string
      name:
        description: 网站名称
        maxLength: 50
        type: string
      screenshot:
        description: 网站截图（可选）
        maxLength: 255
        type: string
      url:
        description: 网站链接
        maxLength: 255
        type: string
    required:
    - avatar
    - description
    - name
    - url
    type: object
...
  dto.CreateFriendRequest:
    properties:
      avatar:
        maxLength: 255
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 50
        type: string
      rss_url:
        description: RSS订阅地址
        maxLength: 500
        type: string
      screenshot:
        maxLength: 255
        type: string
      sort:
        maximum: 10
        minimum: 1
        type: integer
      type_id:
        type: integer
      url:
        maxLength: 255
        type: string
    required:
    - name
    - url
    type: object
  dto.CreateFriendTypeRequest:
    properties:
      is_visible:
        type: boolean
      name:
        maxLength: 50
        type: string
      sort:
        type: integer
    required:
    - name
    type: object
...
  dto.UpdateFriendRequest:
    properties:
      accessible:
        description: 可访问性状态
        type: integer
      avatar:
        maxLength: 255
        type: string
      description:
        maxLength: 500
        type: string
      is_invalid:
        description: 是否失效
        type: boolean
      is_pending:
        description: 是否为待审核申请
        type: boolean
      name:
        maxLength: 50
        type: string
      rss_url:
        description: RSS订阅地址
        maxLength: 500
        type: string
      screenshot:
        maxLength: 255
        type: string
      sort:
        maximum: 10
        minimum: 1
        type: integer
      type_id:
        type: integer
      url:
        maxLength: 255
        type: string
    required:
    - name
    - url
    type: object
  dto.UpdateFriendTypeRequest:
    properties:
      is_visible:
        type: boolean
      name:
        maxLength: 50
        type: string
      sort:
        type: integer
    required:
    - name
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: M.Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Flec-Server
  version: v1
paths:
...
  /admin/friends:
    get:
      consumes:
      - application/json
      description: 获取所有友链用于管理
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量（不传则返回全部）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 友链列表
      tags:
      - 友链管理
    post:
      consumes:
      - application/json
      description: 创建新友链
      parameters:
      - description: 友链信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFriendRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建友链
      tags:
      - 友链管理
  /admin/friends/{id}:
    delete:
      consumes:
      - application/json
      description: 软删除友链
      parameters:
      - description: 友链 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除友链
      tags:
      - 友链管理
    get:
      consumes:
      - application/json
      description: 通过 ID 获取友链信息
      parameters:
      - description: 友链 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 友链详情
      tags:
      - 友链管理
    put:
      consumes:
      - application/json
      description: 修改友链信息
      parameters:
      - description: 友链 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 友链信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新友链
      tags:
      - 友链管理
  /admin/friends/types:
    get:
      description: 获取所有友链类型（包括禁用的）
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量（不传则返回全部）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 友链类型列表 [类型]
      tags:
      - 友链管理
    post:
      consumes:
      - application/json
      description: 创建新的友链类型
      parameters:
      - description: 类型信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFriendTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建友链类型 [类型]
      tags:
      - 友链管理
  /admin/friends/types/{id}:
    delete:
      description: 删除友链类型（关联的友链 type_id 会被设置为 NULL）
      parameters:
      - description: 类型 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除友链类型 [类型]
      tags:
      - 友链管理
    get:
      description: 通过 ID 获取友链类型信息
      parameters:
      - description: 类型 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 友链类型详情 [类型]
      tags:
      - 友链管理
    put:
      consumes:
      - application/json
      description: 修改友链类型信息
      parameters:
      - description: 类型 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 类型信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFriendTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新友链类型 [类型]
      tags:
      - 友链管理
...
  /friends/apply:
    post:
      consumes:
      - application/json
      description: 用户提交友链申请，系统将通知管理员审核（需要登录）
      parameters:
      - description: 友链申请信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApplyFriendRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 申请友链
      tags:
      - 友链
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"